improve Q1_01 efficiency and Ruby idiom
The answer currently posted for Q1_01 works but could be improved in a
few ways. This PR accomplishes the following:

1) makes the algorithm more efficient by using a hash table rather than
Ruby's `String#include?`, which ends up being a loop inside of a loop,
so O(n) instead of O(n^2). It also more closely mirrors the solution in
the book.

2) improves Ruby idiom by:
  - indenting 2 spaces instead of 4
  - ending a predicate method with a question mark
  - using Ruby's `Enumerable#each_with_object` pattern and avoiding
    certain bits of extraneous code in the prior version