Add hash-based option for q1_02
This is actually a larger set of changes than the main line of the
commit message would indicate. This is because I realized it might be
nice not to have a new file for every different approach but to use a
Ruby module with method names that represent what is unique about the
given approach.

The function that I added is very similar to what was called "solution
2" but uses a hash table instead of an array. I think it's worth having
both solutions since the O values are the same but they provide
different approaches. The array is a translation of what is provided in
the book but (imho) the hash version feels like a more idiomatic Ruby
solution (and should reduce space complexity since you don't have to
initialize an Array with 128 chars).

In the course of making the changes I also implemented some other
Ruby-isms that I thought might be useful for users of this repository to
absorb while learning algorithms in Ruby. These include:

- renaming the spec directory `spec/`. This is a naming convention for
  RSpec, and allows the developer to simply run the `$ rspec` command
from the parent directory to run all tests(RSpec will automatically look for a
directory with tests)

- using snake case in filenames (including no capital letters)

- some small stylistic things, like preferring `unless...==` instead of
  `if...!=`, using `string.chars.sort` instead of
`string.split("").sort`, etc.

- put related functions into their own module, so we could have 1 test
  file per exercise and 1 implementation file per exercise

- example of more idiomatic use of `context` and `it` blocks in rspec
  tests to make tests more organized and readable